Added collapsible layout & made toolbar translucent with a gradient
