+++ app/src/main/res/layout/activity_trip.xml
@@ -7,8 +7,6 @@
     android:orientation="vertical"
     android:weightSum="1">
 
-    <include layout="@layout/toolbar_top" />
-
     <FrameLayout
         android:id="@+id/flTrip"
         android:layout_width="match_parent"
