+++ app/src/main/res/layout/fragment_view_pager_container.xml
@@ -1,22 +1,68 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent" android:layout_height="match_parent">
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools">
+    <android.support.design.widget.AppBarLayout
+        android:id="@+id/appbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:fitsSystemWindows="true"
+        android:background="@android:color/transparent"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
+
+    <android.support.design.widget.CollapsingToolbarLayout
+        android:id="@+id/collapsing_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:fitsSystemWindows="true"
+        app:layout_scrollFlags="scroll|exitUntilCollapsed">
+
 
     <ImageView
         android:id="@+id/ivCoverImage"
         android:scaleType="centerCrop"
         android:transitionName="cover"
+        app:layout_collapseMode="parallax"
+        app:layout_collapseParallaxMultiplier="0.9"
+        app:layout_scrollFlags="scroll|enterAlways|enterAlwaysCollapsed"
         android:layout_width="match_parent"
-        android:layout_height="150dp" />
-    <com.astuetz.PagerSlidingTabStrip
-        android:id="@+id/tabStrip"
-        app:pstsShouldExpand="true"
-        android:layout_below="@+id/ivCoverImage"
+        android:layout_height="250dp" />
+
+    <android.support.v7.widget.Toolbar
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/toolbarViewPager"
+        android:minHeight="?attr/actionBarSize"
         android:layout_width="match_parent"
-        android:layout_height="48dp"
-        android:textSize="14sp"
-        android:textColor="#000000" />
+        android:layout_height="wrap_content"
+        app:titleTextColor="@android:color/white"
+        android:background="@drawable/background_toolbar_translucent"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true">
+
+        <TextView
+            android:id="@+id/tvToolbarTitle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            tools:text="San Francisco"
+            android:textColor="@android:color/white"
+            style="@style/TextAppearance.AppCompat.Widget.ActionBar.Title"
+            android:layout_gravity="left" />
+
+    </android.support.v7.widget.Toolbar>
+    </android.support.design.widget.CollapsingToolbarLayout>
+
+        <com.astuetz.PagerSlidingTabStrip
+            android:id="@+id/tabStrip"
+            app:pstsShouldExpand="true"
+            android:layout_below="@+id/ivCoverImage"
+            android:layout_width="match_parent"
+            android:layout_height="48dp"
+            android:textSize="14sp"
+            android:textColor="#000000" />
+    </android.support.design.widget.AppBarLayout>
 
     <android.support.v4.view.ViewPager
         android:id="@+id/viewPager"
@@ -24,6 +70,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@android:color/white"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior"
         android:layout_alignParentBottom="true" />
 
-</RelativeLayout>
+</android.support.design.widget.CoordinatorLayout>
